// src/interface/request/asset.interface.ts
export interface StoreAsset {
  idBarang?: string;
  kodeBarang?: string;
  namaBarang?: string;
  nomorRegistrasi?: string;
  harga?: number;
  merkTipe?: string;
  ukuranCC?: string;
  jenisBahan?: string;
  nomorPabrik?: string;
  nomorRangka?: string;
  nomorMesin?: string;
  nomorPolisi?: string;
  dokumenTipe?: string;
  dokumenNomor?: string;
  keterangan?: string;
  tahunPerolehan?: string;
  kategori?: string;
  dokumen?: string;
}